pyproject.toml
[project]
name = "screen-logger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyobjc-framework-quartz>=12.1",
    "pyobjc-framework-vision>=12.1",
    "pyyaml>=6.0.3",
]
